fix storage descriptor
